[42]: add feature:search module
