+++ app/src/main/java/com/github/tinkzhang/readkeeper/HomeScreen.kt
@@ -12,9 +12,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
+import com.github.tinkzhang.basic.SCREEN_ROUTE
 import com.github.tinkzhang.reading.ReadingViewModel
-import com.github.tinkzhang.readkeeper.ui.SCREEN_ROUTE
-
 
 @Composable
 fun HomeScreen(
