+++ app/src/main/java/com/github/tinkzhang/readkeeper/ui/Navigation.kt
@@ -7,25 +7,27 @@ import androidx.compose.material.icons.outlined.Archive
 import androidx.compose.material.icons.outlined.FavoriteBorder
 import androidx.compose.material.icons.outlined.Home
 import androidx.compose.ui.graphics.vector.ImageVector
+import com.github.tinkzhang.basic.SCREEN_ROUTE
+import com.github.tinkzhang.basic.SCREEN_ROUTE.ARCHIVED_LIST
+import com.github.tinkzhang.basic.SCREEN_ROUTE.HOME
+import com.github.tinkzhang.basic.SCREEN_ROUTE.READING_LIST
+import com.github.tinkzhang.basic.SCREEN_ROUTE.SEARCH
+import com.github.tinkzhang.basic.SCREEN_ROUTE.SETTINGS
+import com.github.tinkzhang.basic.SCREEN_ROUTE.WISH_LIST
 import com.github.tinkzhang.readkeeper.R
-import com.github.tinkzhang.readkeeper.ui.SCREEN_ROUTE.ARCHIVED_LIST
-import com.github.tinkzhang.readkeeper.ui.SCREEN_ROUTE.HOME
-import com.github.tinkzhang.readkeeper.ui.SCREEN_ROUTE.READING_LIST
-import com.github.tinkzhang.readkeeper.ui.SCREEN_ROUTE.SEARCH
-import com.github.tinkzhang.readkeeper.ui.SCREEN_ROUTE.SETTINGS
-import com.github.tinkzhang.readkeeper.ui.SCREEN_ROUTE.WISH_LIST
 
-object SCREEN_ROUTE {
-    const val HOME = "home"
-    const val READING_LIST = "reading_list"
-    const val READING_ITEM =
-        "reading_item/{uuid}?open_progress_dialog={open_progress_dialog}&open_edit_dialog={open_edit_dialog}"
-    const val WISH_LIST = "wish_list"
-    const val ARCHIVED_LIST = "archived_list"
-    const val SEARCH = "search"
-    const val SEARCH_RESUTL = "search_result/{keyword}"
-    const val SETTINGS = "settings"
-}
+
+//object SCREEN_ROUTE {
+//    const val HOME = "home"
+//    const val READING_LIST = "reading_list"
+//    const val READING_ITEM =
+//        "reading_item/{uuid}?open_progress_dialog={open_progress_dialog}&open_edit_dialog={open_edit_dialog}"
+//    const val WISH_LIST = "wish_list"
+//    const val ARCHIVED_LIST = "archived_list"
+//    const val SEARCH = "search"
+//    const val SEARCH_RESUTL = "search_result/{keyword}"
+//    const val SETTINGS = "settings"
+//}
 
 val ROUTE_TO_SCREEN_MAP = mapOf(
     HOME to MainScreenViewData.HomeScreenViewData,
