+++ common/basic/src/main/res/values/strings.xml
@@ -1,4 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="google_sign_in_token_id">594810278400-tvviknir2uqc2qd95uts1l5s54309als.apps.googleusercontent.com</string>
+    <string name="add_wish">+Wish</string>
+    <string name="wished">Added into the wish list</string>
+    <string name="add_reading">+Reading</string>
+    <string name="reading_added">Added into the reading list</string>
 </resources>
