+++ feature/search/src/main/java/com/github/tinkzhang/search/SearchBookItem.kt
@@ -1,4 +1,4 @@
-package com.github.tinkzhang.readkeeper.search
+package com.github.tinkzhang.search
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
@@ -11,11 +11,9 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil.compose.LocalImageLoader
 import coil.compose.rememberImagePainter
-import com.github.tinkzhang.readkeeper.R
 import com.github.tinkzhang.basic.model.SearchBook
 import com.github.tinkzhang.readkeeper.ui.components.ReadingIconToggleButton
 import com.github.tinkzhang.readkeeper.ui.components.WishIconToggleButton
-import com.github.tinkzhang.readkeeper.ui.theme.ReadKeeperTheme
 
 @Composable
 fun SearchBookItem(book: SearchBook) {
@@ -134,7 +132,5 @@ private fun SearchBookItemActionBar(
 @Preview
 @Composable
 fun SearchBookItemActionBarPreview() {
-    ReadKeeperTheme {
-        SearchBookItemActionBar()
-    }
+    SearchBookItemActionBar()
 }
