+++ feature/search/src/main/java/com/github/tinkzhang/search/SearchResultViewModel.kt
@@ -1,4 +1,4 @@
-package com.github.tinkzhang.readkeeper.search
+package com.github.tinkzhang.search
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
@@ -7,12 +7,12 @@ import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.cachedIn
 import androidx.paging.map
+import com.github.tinkzhang.basic.UserRepository
+import com.github.tinkzhang.basic.model.BookInfo
 import com.github.tinkzhang.basic.model.ReadingBook
 import com.github.tinkzhang.basic.model.SearchBook
 import com.github.tinkzhang.basic.model.WishBook
-import com.github.tinkzhang.basic.model.BookInfo
-import com.github.tinkzhang.readkeeper.search.network.SIZE
-import com.github.tinkzhang.basic.UserRepository
+import com.github.tinkzhang.search.network.SIZE
 import github.tinkzhang.readkeeper.search.model.googlebook.Item
 import kotlinx.coroutines.flow.map
