+++ feature/search/src/main/java/com/github/tinkzhang/search/network/googlebook/SearchInfo.kt
